+++ test/end-to-end/offer/:id/archive.js
@@ -1,0 +1,3 @@
+test("bruh", () => {
+    expect(true).toBe(true);
+});
