+++ test/end-to-end/offer/:id/disable.js
@@ -1,0 +1,3 @@
+test("bruh", () => {
+    expect(true).toBe(true);
+});
