+++ test/end-to-end/offer/:id/index.js
@@ -1,0 +1,3 @@
+test("bruh", () => {
+    expect(true).toBe(true);
+});
