+++ test/end-to-end/offer/edit/:offerId/index.js
@@ -1,0 +1,3 @@
+test("bruh", () => {
+    expect(true).toBe(true);
+});
